+++ tortoise/queryset.py
@@ -240,10 +240,16 @@ class AwaitableQuery(Generic[MODEL]):
 
         return any(info["field"].is_aggregate for info in annotation_info.values())
 
-    def sql(self) -> str:
+    def sql(self, **kwargs) -> str:
         """Return the actual SQL."""
+        return self.as_query().get_sql(**kwargs)
+
+    def as_query(self) -> QUERY:
+        """Return the actual query."""
+        if self._db is None:
+            self._db = self._choose_db()  # type: ignore
         self._make_query()
-        return self.query.get_sql()
+        return self.query
 
     def _make_query(self) -> None:
         raise NotImplementedError()  # pragma: nocoverage
@@ -276,6 +282,8 @@ class QuerySet(AwaitableQuery[MODEL]):
         "_select_for_update_of",
         "_select_related",
         "_select_related_idx",
+        "_use_indexes",
+        "_force_indexes",
     )
 
     def __init__(self, model: Type[MODEL]) -> None:
@@ -303,6 +311,8 @@ class QuerySet(AwaitableQuery[MODEL]):
         self._select_related_idx: List[
             Tuple["Type[Model]", int, str, "Type[Model]"]
         ] = []  # format with: model,idx,model_name,parent_model
+        self._force_indexes: Set[str] = set()
+        self._use_indexes: Set[str] = set()
 
     def _clone(self) -> "QuerySet[MODEL]":
         queryset = QuerySet.__new__(QuerySet)
@@ -333,6 +343,8 @@ class QuerySet(AwaitableQuery[MODEL]):
         queryset._select_for_update_of = self._select_for_update_of
         queryset._select_related = self._select_related
         queryset._select_related_idx = self._select_related_idx
+        queryset._force_indexes = self._force_indexes
+        queryset._use_indexes = self._use_indexes
         return queryset
 
     def _filter_or_exclude(self, *args: Q, negate: bool, **kwargs: Any) -> "QuerySet[MODEL]":
@@ -454,7 +466,7 @@ class QuerySet(AwaitableQuery[MODEL]):
             return queryset
         return self
 
-    def annotate(self, **kwargs: Function) -> "QuerySet[MODEL]":
+    def annotate(self, **kwargs: Union[Function, Term]) -> "QuerySet[MODEL]":
         """
         Annotate result with aggregation or function result.
 
@@ -464,8 +476,8 @@ class QuerySet(AwaitableQuery[MODEL]):
 
         queryset = self._clone()
         for key, annotation in kwargs.items():
-            if not isinstance(annotation, (Function, Term)):
-                raise TypeError("value is expected to be Function/Term instance")
+            # if not isinstance(annotation, (Function, Term)):
+            #     raise TypeError("value is expected to be Function/Term instance")
             queryset._annotations[key] = annotation
             queryset._custom_filters.update(get_filters_for_field(key, None, key))
         return queryset
@@ -564,6 +576,8 @@ class QuerySet(AwaitableQuery[MODEL]):
             q_objects=self._q_objects,
             annotations=self._annotations,
             custom_filters=self._custom_filters,
+            limit=self._limit,
+            orderings=self._orderings,
         )
 
     def update(self, **kwargs: Any) -> "UpdateQuery":
@@ -585,6 +599,8 @@ class QuerySet(AwaitableQuery[MODEL]):
             q_objects=self._q_objects,
             annotations=self._annotations,
             custom_filters=self._custom_filters,
+            limit=self._limit,
+            orderings=self._orderings,
         )
 
     def count(self) -> "CountQuery":
@@ -683,6 +699,29 @@ class QuerySet(AwaitableQuery[MODEL]):
             queryset._select_related.add(field)
         return queryset
 
+    def force_index(self, *index_names: str) -> "QuerySet[MODEL]":
+        """
+        The FORCE INDEX hint acts like USE INDEX (index_list),
+        with the addition that a table scan is assumed to be very expensive.
+        """
+        if self.capabilities.support_index_hint:
+            queryset = self._clone()
+            for index_name in index_names:
+                queryset._force_indexes.add(index_name)
+            return queryset
+        return self
+
+    def use_index(self, *index_names: str) -> "QuerySet[MODEL]":
+        """
+        The USE INDEX (index_list) hint tells MySQL to use only one of the named indexes to find rows in the table.
+        """
+        if self.capabilities.support_index_hint:
+            queryset = self._clone()
+            for index_name in index_names:
+                queryset._use_indexes.add(index_name)
+            return queryset
+        return self
+
     def prefetch_related(self, *args: Union[str, Prefetch]) -> "QuerySet[MODEL]":
         """
         Like ``.fetch_related()`` on instance, but works on all objects in QuerySet.
@@ -756,9 +795,9 @@ class QuerySet(AwaitableQuery[MODEL]):
 
         table = self._join_table_by_field(table, field, field_object)
         related_fields = field_object.related_model._meta.db_fields
-        self._select_related_idx.append(
-            (field_object.related_model, len(related_fields), field, model)
-        )
+        append_item = (field_object.related_model, len(related_fields), field, model)
+        if append_item not in self._select_related_idx:
+            self._select_related_idx.append(append_item)
         for related_field in related_fields:
             self.query = self.query.select(
                 table[related_field].as_(f"{table.get_table_name()}.{related_field}")
@@ -778,9 +817,9 @@ class QuerySet(AwaitableQuery[MODEL]):
         self._select_related_idx = []
         table = self.model._meta.basetable
         if self._fields_for_select:
-            self._select_related_idx.append(
-                (self.model, len(self._fields_for_select), table, self.model)
-            )
+            append_item = (self.model, len(self._fields_for_select), table, self.model)
+            if append_item not in self._select_related_idx:
+                self._select_related_idx.append(append_item)
             db_fields_for_select = [
                 table[self.model._meta.fields_db_projection[field]].as_(field)
                 for field in self._fields_for_select
@@ -788,9 +827,9 @@ class QuerySet(AwaitableQuery[MODEL]):
             self.query = copy(self.model._meta.basequery).select(*db_fields_for_select)
         else:
             self.query = copy(self.model._meta.basequery_all_fields)
-            self._select_related_idx.append(
-                (self.model, len(self.model._meta.db_fields), table, self.model)
-            )
+            append_item = (self.model, len(self.model._meta.db_fields), table, self.model)
+            if append_item not in self._select_related_idx:
+                self._select_related_idx.append(append_item)
         self.resolve_ordering(
             self.model, self.model._meta.basetable, self._orderings, self._annotations
         )
@@ -821,6 +860,10 @@ class QuerySet(AwaitableQuery[MODEL]):
                     field=field_split[0],
                     forwarded_fields="__".join(field_split[1:]) if len(field_split) > 1 else "",
                 )
+        if self._force_indexes:
+            self.query = self.query.force_index(*self._force_indexes)
+        if self._use_indexes:
+            self.query = self.query.use_index(*self._use_indexes)
 
     def __await__(self) -> Generator[Any, None, List[MODEL]]:
         if self._db is None:
@@ -852,7 +895,14 @@ class QuerySet(AwaitableQuery[MODEL]):
 
 
 class UpdateQuery(AwaitableQuery):
-    __slots__ = ("update_kwargs", "q_objects", "annotations", "custom_filters")
+    __slots__ = (
+        "update_kwargs",
+        "q_objects",
+        "annotations",
+        "custom_filters",
+        "orderings",
+        "limit",
+    )
 
     def __init__(
         self,
@@ -862,6 +912,8 @@ class UpdateQuery(AwaitableQuery):
         q_objects: List[Q],
         annotations: Dict[str, Any],
         custom_filters: Dict[str, Dict[str, Any]],
+        limit: Optional[int],
+        orderings: List[Tuple[str, str]],
     ) -> None:
         super().__init__(model)
         self.update_kwargs = update_kwargs
@@ -869,10 +921,16 @@ class UpdateQuery(AwaitableQuery):
         self.annotations = annotations
         self.custom_filters = custom_filters
         self._db = db
+        self.limit = limit
+        self.orderings = orderings
 
     def _make_query(self) -> None:
         table = self.model._meta.basetable
         self.query = self._db.query_class.update(table)
+        if self.capabilities.support_update_limit_order_by and self.limit:
+            self.query._limit = self.limit
+            self.resolve_ordering(self.model, table, self.orderings, self.annotations)
+
         self.resolve_filters(
             model=self.model,
             q_objects=self.q_objects,
@@ -919,7 +977,7 @@ class UpdateQuery(AwaitableQuery):
 
 
 class DeleteQuery(AwaitableQuery):
-    __slots__ = ("q_objects", "annotations", "custom_filters")
+    __slots__ = ("q_objects", "annotations", "custom_filters", "orderings", "limit")
 
     def __init__(
         self,
@@ -928,15 +986,24 @@ class DeleteQuery(AwaitableQuery):
         q_objects: List[Q],
         annotations: Dict[str, Any],
         custom_filters: Dict[str, Dict[str, Any]],
+        limit: Optional[int],
+        orderings: List[Tuple[str, str]],
     ) -> None:
         super().__init__(model)
         self.q_objects = q_objects
         self.annotations = annotations
         self.custom_filters = custom_filters
         self._db = db
+        self.limit = limit
+        self.orderings = orderings
 
     def _make_query(self) -> None:
         self.query = copy(self.model._meta.basequery)
+        if self.capabilities.support_update_limit_order_by and self.limit:
+            self.query._limit = self.limit
+            self.resolve_ordering(
+                self.model, self.model._meta.basetable, self.orderings, self.annotations
+            )
         self.resolve_filters(
             model=self.model,
             q_objects=self.q_objects,