+++ CHANGELOG.rst
@@ -9,6 +9,41 @@ Changelog
 
 0.17
 ====
+0.17.5
+------
+- Set `field_type` of fk and o2o same to which relation field type. (#443)
+
+0.17.4
+------
+- Fix `update_or_create`. (#782)
+
+0.17.3
+------
+- Fix duplicates when using custom through association class on M2M relations
+- Fix `update_or_create` and `get_or_create`. (#721)
+- Fix `refresh_from_db` without fields pass. (#734)
+- Make `update` query work with `limit` and `order_by`. (#748)
+- Add `Subquery` expression. (#756) (#9) (#337)
+- Use JSON in JSONField.
+
+0.17.2
+------
+- Add more `index` types.
+- Add `force_index`, `use_index` to `queryset`.
+- Fix `F` in update error with `update_fields`.
+- Make `delete` query work with `limit` and `order_by`. (#697)
+- Filter backward FK fields with `IS NULL` and `NOT IS NULL` filters (#700)
+- Add `select_for_update` in `update_or_create`. (#702)
+- Add `Model.select_for_update`.
+- Add `__search` full text search to queryset.
+
+0.17.1
+------
+- Fix type for modules.
+- Fix `select_related` when related model specified more than once. (#679)
+- Add `__iter__` to model, now can just return model/models in `fastapi` response.
+- Fix `in_transaction` bug caused by 'router'. (#677) (#678)
+
 0.17.0
 -------
 - Add date part extract filtering.
@@ -17,6 +52,7 @@ Changelog
 - Add `nowait`, `skip_locked`, `of` parameters to `queryset.select_for_update`.
 - Add field name to validation exceptions.
 - Compatible with `asyncmy <https://github.com/long2ice/asyncmy>`_.
+- Replace pypika to `pypika-tortoise <https://github.com/tortoise/pypika-tortoise>`_.
 
 0.16
 ====